requirements.txt
lyzr_automata
solara
python-dotenv